🔥 용도를 다한 mock데이터 제거 및 디버깅용 console.log 모두 제거
