+++ src/components/common/GNB.tsx
@@ -2,15 +2,12 @@
 import React, { useEffect, useRef, useState } from "react";
 import Link from "next/link";
 import NotificationModalComponent from "./NotificationModal/NotificationModalComponent";
-import { MockData } from "./NotificationModal/NOTIFICATION_API_RESPONSE_TYPE";
+
 import SearchSvg from "./GNB/SearchSvg";
 import LogoSvg from "./GNB/LogoSvg";
 import { getCookie } from "@/util/cookieSetting";
 import { usePathname, useRouter } from "next/navigation";
 import { logout, mydataApiResponse } from "@/util/api";
-import { NOTIFICATION_API_RESPONSE_TYPE } from "./NotificationModal/NOTIFICATION_API_RESPONSE_TYPE";
-import { NOTIFICATION_API_ITEM_TYPE } from "./NotificationModal/NOTIFICATION_API_RESPONSE_TYPE";
-import { alertApiResponse } from "@/util/api";
 
 //TODO: 유져의 알림 목록 조회 API 연동 필요
