+++ src/components/common/NoticeDetail/NoticeDetailButtons.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React, { useEffect, useState } from "react";
 import Button from "../Button";
-import { usePathname, useRouter } from "next/navigation";
+import { useRouter } from "next/navigation";
 import {
   mydataApiResponse,
   searchUserApplyApiResponse,
