--- src/util/constants/EMPLOYER_POST_LIST.ts
@@ -1,348 +1,0 @@
-const EMPLOYER_POST_LIST = {
-  offset: 0,
-  limit: 10,
-  count: 10,
-  hasNext: false,
-  items: [
-    {
-      item: {
-        id: "6fa67d08-de36-4b13-b085-e530d43f65a3",
-        hourlyPay: 500000,
-        startsAt: "2024-05-01T00:00:00.000Z",
-        workhour: 5,
-        description: "string",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/6fa67d08-de36-4b13-b085-e530d43f65a3",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/6fa67d08-de36-4b13-b085-e530d43f65a3",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "9edbf970-d9a0-4504-99c5-497063c9e0b5",
-        hourlyPay: 50000,
-        startsAt: "2024-05-01T00:00:00.000Z",
-        workhour: 5,
-        description: "string",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/9edbf970-d9a0-4504-99c5-497063c9e0b5",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/9edbf970-d9a0-4504-99c5-497063c9e0b5",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "4539ab39-d627-4343-8f4d-27c4f5048d12",
-        hourlyPay: 50000,
-        startsAt: "2024-05-01T00:00:00.000Z",
-        workhour: 5,
-        description: "string",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/4539ab39-d627-4343-8f4d-27c4f5048d12",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/4539ab39-d627-4343-8f4d-27c4f5048d12",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "67646731-b41e-4318-a5de-1b9c78e22ca8",
-        hourlyPay: 50000,
-        startsAt: "2024-05-01T00:00:00.000Z",
-        workhour: 5,
-        description: "string",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/67646731-b41e-4318-a5de-1b9c78e22ca8",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/67646731-b41e-4318-a5de-1b9c78e22ca8",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "ba1e7554-0b2b-4583-9a6f-a16dae71b46a",
-        hourlyPay: 500000,
-        startsAt: "2024-05-01T00:00:00.000Z",
-        workhour: 5,
-        description: "string",
-        closed: true,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/ba1e7554-0b2b-4583-9a6f-a16dae71b46a",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/ba1e7554-0b2b-4583-9a6f-a16dae71b46a",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "a9989e2e-253c-40ce-8ecd-6f946da626e0",
-        hourlyPay: 50000,
-        startsAt: "2024-05-01T00:00:00.000Z",
-        workhour: 5,
-        description: "string",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/a9989e2e-253c-40ce-8ecd-6f946da626e0",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/a9989e2e-253c-40ce-8ecd-6f946da626e0",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "7f642440-619b-4940-95bf-90905d6498b3",
-        hourlyPay: 100002,
-        startsAt: "2024-04-29T00:00:00Z",
-        workhour: 10,
-        description: "string",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/7f642440-619b-4940-95bf-90905d6498b3",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/7f642440-619b-4940-95bf-90905d6498b3",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "c43c080d-e036-42d0-bebd-0f62b978e04c",
-        hourlyPay: 24000,
-        startsAt: "2024-04-26T15:00:00.000Z",
-        workhour: 24,
-        description: "ㅂㅂ",
-        closed: true,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/c43c080d-e036-42d0-bebd-0f62b978e04c",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/c43c080d-e036-42d0-bebd-0f62b978e04c",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "1b12e4aa-2aeb-472a-919b-eaf102b6e7c6",
-        hourlyPay: 20000,
-        startsAt: "2024-04-18T15:00:00.000Z",
-        workhour: 12,
-        description: "주말 알바 시급",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/1b12e4aa-2aeb-472a-919b-eaf102b6e7c6",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/1b12e4aa-2aeb-472a-919b-eaf102b6e7c6",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-    {
-      item: {
-        id: "b0dc8d8a-ab60-4b99-98af-95e5bbeefbea",
-        hourlyPay: 15000,
-        startsAt: "2024-04-12T15:00:00.000Z",
-        workhour: 12,
-        description: "시간은 길지만 편해유",
-        closed: false,
-      },
-      links: [
-        {
-          rel: "self",
-          description: "공고 정보",
-          method: "GET",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/b0dc8d8a-ab60-4b99-98af-95e5bbeefbea",
-        },
-        {
-          rel: "update",
-          description: "공고 수정",
-          method: "PUT",
-          href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices/b0dc8d8a-ab60-4b99-98af-95e5bbeefbea",
-          body: {
-            hourlyPay: "number",
-            startsAt: "string",
-            workhour: "number",
-            description: "string",
-          },
-        },
-      ],
-    },
-  ],
-  links: [
-    {
-      ref: "self",
-      description: "현재 페이지",
-      method: "GET",
-      href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices?offset=0&limit=10",
-    },
-    {
-      rel: "prev",
-      description: "이전 페이지",
-      method: "GET",
-      href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices?offset=0&limit=10",
-    },
-    {
-      rel: "next",
-      description: "다음 페이지",
-      method: "GET",
-      href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices?offset=10&limit=10",
-    },
-    {
-      rel: "create",
-      description: "공고 생성",
-      method: "POST",
-      href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57/notices",
-      body: {
-        hourlyPay: "number",
-        startsAt: "string",
-        workhour: "number",
-        description: "string",
-      },
-    },
-    {
-      rel: "shop",
-      description: "가게 정보",
-      method: "GET",
-      href: "/api/0-1/the-julge/shops/52200d57-25a3-4910-9903-e524d1d67a57",
-    },
-  ],
-};
-
-export default EMPLOYER_POST_LIST;
