--- src/util/constants/NOTIFICATION_MODAL_MOCK_DATA.ts
@@ -1,65 +1,0 @@
-const Data = [
-    {
-        state : true,
-        message : "This is a notification message1",
-        time : "20",
-    }, 
-    {
-        state : false,
-        message : "This is a notification message2",
-        time : "30",
-
-    }, 
-    {
-        state : true,
-        message : "This is a notification message3",
-        time : "40",
-    }, 
-    {
-        state : false,
-        message : "This is a notification message4",
-        time : "50",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-    {
-        state : true,
-        message : "This is a notification message5",
-        time : "60",
-    }, 
-]
-
-export default Data;
